Code tidy up and xml docs
