+++ src/CommunityToolkit/Xamarin.CommunityToolkit/Behaviors/Animations/AnimationTypes/AnimationBase.shared.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Xamarin.Forms;
 
 namespace Xamarin.CommunityToolkit.Behaviors
@@ -39,92 +36,5 @@ namespace Xamarin.CommunityToolkit.Behaviors
 
 	public abstract class AnimationBase : AnimationBase<View>
 	{
-	}
-
-	public abstract class PreBuiltAnimationBase : AnimationBase
-	{
-		public static readonly BindableProperty IsRepeatedProperty =
-		   BindableProperty.Create(nameof(IsRepeated), typeof(bool), typeof(PreBuiltAnimationBase), default, BindingMode.TwoWay);
-
-		public bool IsRepeated
-		{
-			get => (bool)GetValue(IsRepeatedProperty);
-			set => SetValue(IsRepeatedProperty, value);
-		}
-
-		public override Task Animate(View? view) => Animate(CancellationToken.None, view!);
-
-		public async Task Animate(CancellationToken? cancellationToken, params View[]? views)
-		{
-			if (views != null)
-			{
-				AnimationWrapper? animation = null;
-
-				var taskCompletionSource = new TaskCompletionSource<bool>();
-
-				if (cancellationToken is null)
-					cancellationToken = CancellationToken.None;
-
-				animation = CreateAnimation(
-					16,
-					onFinished: (v, c) =>
-					{
-						try
-						{
-							if (cancellationToken != null)
-							{
-								cancellationToken.Value.ThrowIfCancellationRequested();
-							}
-
-							if (IsRepeated)
-							{
-								return;
-							}
-
-							animation = null;
-							taskCompletionSource.SetResult(c);
-						}
-						catch (OperationCanceledException)
-						{
-							taskCompletionSource.SetCanceled();
-							animation?.Abort();
-						}
-					},
-					shouldRepeat: () => IsRepeated,
-					views);
-
-				animation.Commit();
-
-				await Task.WhenAny(cancellationToken.Value.WhenCanceled(), taskCompletionSource.Task);
-				animation?.Abort();
-			}
-		}
-
-		AnimationWrapper CreateAnimation(
-			uint rate = 16,
-			Action<double, bool>? onFinished = null,
-			Func<bool>? shouldRepeat = null,
-			params View[] views) =>
-			new AnimationWrapper(
-				CreateAnimation(views),
-				Guid.NewGuid().ToString(),
-				views.First(),
-				rate,
-				Duration,
-				Easing,
-				onFinished,
-				shouldRepeat);
-
-		protected abstract Animation CreateAnimation(params View[] views);
-	}
-
-	public static class TaskExtensions
-	{
-		public static Task WhenCanceled(this CancellationToken cancellationToken)
-		{
-			var tcs = new TaskCompletionSource<bool>();
-			cancellationToken.Register(s => ((TaskCompletionSource<bool>)s).SetResult(true), tcs);
-			return tcs.Task;
-		}
 	}
 }
