+++ src/CommunityToolkit/Xamarin.CommunityToolkit/Behaviors/Animations/AnimationTypes/TadaAnimation.shared.cs
@@ -2,37 +2,65 @@
 
 namespace Xamarin.CommunityToolkit.Behaviors
 {
+	/// <summary>
+	/// A 'Tada' animation. Results in:
+	/// <list type="bullet">
+	/// <item>reducing the scale of the view(s)</item>
+	/// <item>wobbling the view(s) while scaling up</item>
+	/// <item>scaling back down to 100%</item>
+	/// </list>
+	/// </summary>
 	public class TadaAnimation : PreBuiltAnimationBase
 	{
+		/// <summary>
+		/// The <see cref="BindableProperty"/> backing store for the <see cref="MaximumScale"/> property.
+		/// </summary>
 		public static readonly BindableProperty MaximumScaleProperty =
 		   BindableProperty.Create(nameof(MaximumScale), typeof(double), typeof(TadaAnimation), 1.1, BindingMode.TwoWay);
 
+		/// <summary>
+		/// Gets or sets the maximum value the views will be scaled to during the animation.
+		/// </summary>
 		public double MaximumScale
 		{
 			get => (double)GetValue(MaximumScaleProperty);
 			set => SetValue(MaximumScaleProperty, value);
 		}
 
+		/// <summary>
+		/// The <see cref="BindableProperty"/> backing store for the <see cref="MinimumScale"/> property.
+		/// </summary>
 		public static readonly BindableProperty MinimumScaleProperty =
 		   BindableProperty.Create(nameof(MinimumScale), typeof(double), typeof(TadaAnimation), 0.9, BindingMode.TwoWay);
 
+		/// <summary>
+		/// Gets or sets the minimum value the views will be scaled to during the animation.
+		/// </summary>
 		public double MinimumScale
 		{
 			get => (double)GetValue(MinimumScaleProperty);
 			set => SetValue(MinimumScaleProperty, value);
 		}
 
+		/// <summary>
+		/// The <see cref="BindableProperty"/> backing store for the <see cref="RotationAngle"/> property.
+		/// </summary>
 		public static readonly BindableProperty RotationAngleProperty =
 		   BindableProperty.Create(nameof(RotationAngle), typeof(double), typeof(TadaAnimation), 3.0, BindingMode.TwoWay);
 
+		/// <summary>
+		/// Gets or sets the angle the views will be rotated by during the animation.
+		/// </summary>
 		public double RotationAngle
 		{
 			get => (double)GetValue(RotationAngleProperty);
 			set => SetValue(RotationAngleProperty, value);
 		}
 
+		/// <inheritdoc />
 		protected override uint DefaultDuration { get; set; } = 1000;
 
+		/// <inheritdoc />
 		protected override Animation CreateAnimation(params View[] views) => Create(RotationAngle, MinimumScale, MaximumScale, views);
 
 		static Animation Create(double rotationAngle, double minimumScale, double maximumScale, params View[] views)
