+++ src/CommunityToolkit/Xamarin.CommunityToolkit/Xamarin.CommunityToolkit.csproj
@@ -213,5 +213,13 @@
     <Compile Remove="build\*.cs" />
     <None Include="build\**\*.cs;build\**\*.targets" Pack="true" PackagePath="build" />
   </ItemGroup>
+  <ItemGroup>
+    <Compile Update="Extensions\TaskExtensions.shared.cs">
+      <SubType></SubType>
+    </Compile>
+    <Compile Update="Behaviors\Animations\AnimationTypes\PreBuiltAnimationBase.shared.cs">
+      <SubType></SubType>
+    </Compile>
+  </ItemGroup>
   <Import Project="..\..\mdoc.targets" Condition=" '$(OS)' == 'Windows_NT' AND '$(GenerateApiDocs)' == 'true'" />
 </Project>
